made vision align more reusable for next year
