+++ src/main/java/org/usfirst/frc4904/robot/commands/vision/VisionMoveHighPort.java
@@ -1,14 +1,14 @@
 package org.usfirst.frc4904.robot.commands.vision;
 
 import org.usfirst.frc4904.robot.RobotMap;
-import org.usfirst.frc4904.robot.VisionTargetHub.VisionTargetType;
+import org.usfirst.frc4904.robot.VisionTargetType;
 
 /**
  * Move towards the closest high port (where you shoot)
  */
 public class VisionMoveHighPort extends VisionMove {
     public VisionMoveHighPort() {
-        super(RobotMap.Component.hub, VisionTargetType.HIGH_PORT, RobotMap.Component.splinesChassis,
+        super(RobotMap.Component.hub, VisionTargetType.HIGH_PORT.targetType, RobotMap.Component.splinesChassis,
                 RobotMap.Metrics.Chassis.DISTANCE_FRONT_BACK);
     }
 }
