+++ src/main/java/org/usfirst/frc4904/robot/commands/vision/VisionTurn.java
@@ -1,7 +1,6 @@
 package org.usfirst.frc4904.robot.commands.vision;
 
 import org.usfirst.frc4904.robot.VisionTargetHub;
-import org.usfirst.frc4904.robot.VisionTargetHub.VisionTargetType;
 import org.usfirst.frc4904.standard.commands.chassis.ChassisTurn;
 import org.usfirst.frc4904.standard.custom.motioncontrollers.MotionController;
 import org.usfirst.frc4904.standard.custom.sensors.IMU;
@@ -13,7 +12,7 @@ import org.usfirst.frc4904.standard.subsystems.chassis.Chassis;
  */
 public class VisionTurn extends ChassisTurn {
 
-    public VisionTurn(String name, Chassis chassis, VisionTargetHub hub, VisionTargetType intendedTargetType, IMU imu,
+    public VisionTurn(String name, Chassis chassis, VisionTargetHub hub, double intendedTargetType, IMU imu,
             MotionController motionController) {
         super(name, chassis, hub.findIdealTarget(intendedTargetType).theta, imu, motionController);
     }
