+++ src/main/java/org/usfirst/frc4904/robot/commands/vision/VisionTurnLoadingPort.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc4904.robot.commands.vision;
 
 import org.usfirst.frc4904.robot.RobotMap;
-import org.usfirst.frc4904.robot.VisionTargetHub.VisionTargetType;
+import org.usfirst.frc4904.robot.VisionTargetType;
 
 /**
  * Turn the robot in the direction of the nearest loading port (human player
@@ -9,7 +9,7 @@ import org.usfirst.frc4904.robot.VisionTargetHub.VisionTargetType;
  */
 public class VisionTurnLoadingPort extends VisionTurn {
     public VisionTurnLoadingPort() {
-        super("VisionTurnLowPort", RobotMap.Component.chassis, RobotMap.Component.hub, VisionTargetType.LOADING_PORT,
+        super("VisionTurnLowPort", RobotMap.Component.chassis, RobotMap.Component.hub, VisionTargetType.LOADING_PORT.targetType,
                 RobotMap.Component.navx, RobotMap.Component.turnPID);
     }
 }
